requirements.txt
Flask
mysql-connector-python
requests
flask_mysqldb